removed dictionary.py and moved the etymology portion to etymology.py
